+++ bgstally/bgstally.py
@@ -264,6 +264,10 @@ class BGSTally:
                 activity.mission_failed(entry, self.mission_log)
                 dirty = True
 
+            case 'PowerplayMerits':
+                activity.powerplay_merits(entry)
+                dirty = True
+
             case 'ReceiveText':
                 self.target_manager.received_text(entry, system)